requirements.txt
markdown-it-py>=2.2.0
mdit-py-plugins>=0.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0